Rewired Conversational Controller
Removed the "Patron" class as CC had already overtaken all of its functionality. May need to split it back up when we get to verbalizing the underlying data structures, but we can burn that bridge when we get there.

Also went ahead and rewrote where the actions the player can ask about are coming from. Initially it was set up that they came from the patron's history, but now they come from the players kb. Will need to consider how the filtering on which events player can ask about are done. May need to rethink UI.