Added All the Notebook Menu Tabs I had planned
Also I think I've fixed the looping/manipulative nature of NPCs goals (specifically the wanting other people to have items so they could ask for them, they can no longer act on the precondition to actions involving other NPCs inventories. This fixes this situation but might be a bandaid and I discover later this is actually a simptom of a deeper problem, we'll see...)

Also also, added a question confirming whether something is still a given NPCs goal (need to put some time gates on it though so the question doesn't come up for goals we found out about in the same conversation).