Fixed? Playthrough Logging
Centralized all Logging code.
Moved write location from Application.dataPath to Application.persistentDataPath because it sounds like its better for Macs.
Removed the time stamp on the folder as it there is a noticable delay between each set of files and I wanted them all to go to the same folder.